third_party/blink/web_tests/fast/dom/shadow/sibling-rules-under-shadow-root-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
crbug.com/337573: Test direct/indirect adjucent and positinal rules under shadow root.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS colorOf(getNodeInComposedTree("host/child2")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child4")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is not "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
